// Tasks/OrchestratorV1/interfaces.ts
export enum ReleaseStatus {

    Undefined = "Undefined",
    InProgress = "InProgress",
    Succeeded = "Succeeded",
    PartiallySucceeded = "PartiallySucceeded",
    Failed = "Failed",

}
// ...
export interface IReleaseDetails {

    endpointName: string;
    projectName: string;
    releaseName: string;
    requesterName: string;
    requesterId: string;

}

export interface IReleaseParameters {

    projectName: string;
    releaseId: number;
    releaseStages: string[];
    sleep: number;

}

export interface IApproveParameters {

    projectName: string;
    status: IStageApproval;
    retry: number;
    sleep: number;

}

export interface IStageApproval {

    status: ri.ApprovalStatus;
    count: number;

}
// ...
export interface IReleaseFilter {

    artifactVersion?: string;
    sourceBranch?: string;
    tag?: string[];

}

export interface IHelper {

    getProject(projectId: string): Promise<ci.TeamProject>;
    getDefinition(projectName: string, definitionId: number): Promise<ri.ReleaseDefinition>;
    getRelease(project: ci.TeamProject, releaseId: number, stages: string[]): Promise<ri.Release>;
    findRelease(projectName: string, definitionId: number, stages: string[], filter: IReleaseFilter): Promise<ri.Release>;
    createRelease(project: ci.TeamProject, definition: ri.ReleaseDefinition, details: IReleaseDetails, stages?: string[], artifacts?: ri.ArtifactMetadata[]): Promise<ri.Release>;
    findBuild(projectName: string, definitionId: number, tags?: string[]): Promise<bi.Build>;
    updateStage(status: ri.ReleaseEnvironmentUpdateMetadata, projectName: string, releaseId: number, stageId: number): Promise<ri.ReleaseEnvironment>;
    updateApproval(status: ri.ReleaseApproval, projectName: string, requestId: number): Promise<ri.ReleaseApproval>;
    getReleaseStatus(projectName: string, releaseId: number): Promise<ri.Release>;
    getStageStatus(releaseStatus: ri.Release, stageName: string): Promise<ri.ReleaseEnvironment>;
    getArtifacts(projectName: string, definitionId: number, primaryId: string, versionId?: string, sourceBranch?: string): Promise<ri.ArtifactMetadata[]>;
    isAutomated(release: ri.Release): Promise<boolean>;

}

export interface IDeployer {

    deployManual(parameters: IReleaseParameters, releaseDetails: IReleaseDetails): Promise<void>;
    deployAutomated(parameters: IReleaseParameters, releaseDetails: IReleaseDetails): Promise<void>;
    approveStage(stage: ri.ReleaseEnvironment, parameters: IApproveParameters, releaseDetails: IReleaseDetails): Promise<IStageApproval>;
}